requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.3
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
databricks-sdk==0.85.0
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docker==7.1.0
executing==2.2.1
fastapi==0.128.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.1
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
multidict==6.7.1
narwhals==2.16.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.4.2
scipy==1.17.0
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
streamlit==1.54.0
tenacity==9.1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.5.3
websockets==15.0.1
Werkzeug==3.1.5
yarl==1.22.0
zipp==3.23.0